1) Cleaned up implementation of history function. 2) Improved history file format to collect control information (like    steering etc) to allow replaying based on the physics (and not only    replaying recorded kart positions). 3) Added some constants (time till finish, time for music credits) to    the stk_config.dat file. 4) Made name of the menu background picture configureable in    stk_config.dat file.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/trunk/supertuxkart@2328 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/track.cpp
@@ -444,16 +444,17 @@ btTransform Track::getStartTransform(unsigned int pos) const {
 //-----------------------------------------------------------------------------
 /** Determines if a kart moving from sector OLDSEC to sector NEWSEC
  *  would be taking a shortcut, i.e. if the distance is larger
- *  than a certain detla
+ *  than a certain delta
  */
 bool Track::isShortcut(const int OLDSEC, const int NEWSEC) const
 {
     // If the kart was off the road, don't do any shortcuts
     if(OLDSEC==UNKNOWN_SECTOR || NEWSEC==UNKNOWN_SECTOR) return false;
-    int distance_sectors = m_distance_from_start[std::max(NEWSEC, OLDSEC)] - m_distance_from_start[std::min(NEWSEC, OLDSEC)];
+    int distance_sectors = (int)(m_distance_from_start[std::max(NEWSEC, OLDSEC)] -
+                                 m_distance_from_start[std::min(NEWSEC, OLDSEC)]  );
     
     // Handle 'warp around'
-    const int track_length = m_distance_from_start[m_driveline.size()-1];
+    const int track_length = (int)m_distance_from_start[m_driveline.size()-1];
     if( distance_sectors < 0 ) distance_sectors += track_length;
     else if( distance_sectors > track_length/2) distance_sectors -= track_length;
